color2/iColoriT/iColoriT_demo/requirements.txt
torch
torchvision
# torch==1.8.2+cu111
# torchvision==0.9.2+cu111
# pip3 install torch==1.8.2 torchvision==0.9.2 --extra-index-url https://download.pytorch.org/whl/lts/1.8/cu111
einops==0.4.1
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.2.0
tensorboardX==2.5.1
timm==0.4.12
tqdm==4.64.0
PyQt5==5.15.7
scikit_image==0.19.3
#skimage==0.0
